backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pymongo==4.6.0
redis==5.0.1

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Data Processing
pandas==2.1.2
numpy==1.26.1
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
tensorflow==2.14.0

# Async Tasks
celery==5.3.4
flower==2.0.1

# Testing
pytest==7.4.3
httpx==0.25.1

# Utilities
requests==2.31.0
beautifulsoup4==4.12.2
aiohttp==3.8.6
kafka-python==2.0.2 